Integration test fix

+++ src/main/java/com/fbc/bot/service/UserService.java
@@ -15,5 +15,5 @@ public interface UserService {
 
     UserDto getUserById(Long id);
 
-    Optional<User> getUserByTelegramId(Integer telegramId);
+    Optional<User> getUserByTelegramId(Long telegramId);
 }
